Update http controllers
