+++ signature-server/internal/controller/http/v1/license.go
@@ -8,7 +8,7 @@ import (
 	"github.com/valyala/fasthttp"
 )
 
-func (controller *V1) licenseIssue(ctx *fasthttp.RequestCtx) {
+func (controller *V1) issue(ctx *fasthttp.RequestCtx) {
 	var req dto.LicenseIssueRequest
 	if err := httputil.DecodeJSON(ctx, &req); err != nil {
 		httputil.RespondError(ctx, fasthttp.StatusBadRequest, err)
@@ -18,6 +18,7 @@ func (controller *V1) licenseIssue(ctx *fasthttp.RequestCtx) {
 
 	resp, err := controller.licenseUc.Issue(context.Background(), req)
 	if err != nil {
+		controller.l.Error("issue: internal error: %v", err)
 		httputil.RespondError(ctx, fasthttp.StatusInternalServerError, err)
 
 		return
@@ -26,7 +27,7 @@ func (controller *V1) licenseIssue(ctx *fasthttp.RequestCtx) {
 	httputil.RespondJSON(ctx, resp)
 }
 
-func (controller *V1) licenseVerify(ctx *fasthttp.RequestCtx) {
+func (controller *V1) verify(ctx *fasthttp.RequestCtx) {
 	var req dto.LicenseVerifyRequest
 	if err := httputil.DecodeJSON(ctx, &req); err != nil {
 		httputil.RespondError(ctx, fasthttp.StatusBadRequest, err)
@@ -36,10 +37,32 @@ func (controller *V1) licenseVerify(ctx *fasthttp.RequestCtx) {
 
 	resp, err := controller.licenseUc.Verify(context.Background(), req)
 	if err != nil {
+		controller.l.Error("verify: internal error: %v", err)
 		httputil.RespondError(ctx, fasthttp.StatusInternalServerError, err)
 
 		return
 	}
 
-	httputil.RespondJSON(ctx, resp)
+	ctx.SetContentType("text/plain")
+	ctx.Response.SetBodyString(resp.Signature)
+}
+
+func (controller *V1) getPublicKey(ctx *fasthttp.RequestCtx) {
+	publicKey, err := controller.licenseUc.GetPublicKey()
+	if err != nil {
+		controller.l.Error("getPublicKey: failed to get public key: %v", err)
+		httputil.RespondError(ctx, fasthttp.StatusInternalServerError, err)
+
+		return
+	}
+
+	ctx.SetContentType("application/x-pem-file")
+	ctx.SetStatusCode(fasthttp.StatusOK)
+
+	if _, err := ctx.Write(publicKey); err != nil {
+		controller.l.Error("getPublicKey: failed to write response: %v", err)
+		httputil.RespondError(ctx, fasthttp.StatusInternalServerError, err)
+
+		return
+	}
 }
